--- cli/template/addons/trpc/app-router.ts
@@ -1,9 +1,0 @@
-import { router } from "../trpc";
-import { exampleRouter } from "./example";
-
-export const appRouter = router({
-  example: exampleRouter,
-});
-
-// export type definition of API
-export type AppRouter = typeof appRouter;
