--- cli/template/addons/trpc/server-utils.ts
@@ -1,15 +1,0 @@
-import { initTRPC } from "@trpc/server";
-import superjson from "superjson";
-
-import { type Context } from "./context";
-
-const t = initTRPC.context<Context>().create({
-  transformer: superjson,
-  errorFormatter({ shape }) {
-    return shape;
-  },
-});
-
-export const router = t.router;
-
-export const publicProcedure = t.procedure;
