+++ cli/template/extras/config/prettier.config.cjs
