+++ cli/template/extras/config/tailwind.config.cjs
