+++ cli/template/extras/src/pages/_app/with-auth.tsx
