+++ cli/template/extras/src/pages/_app/with-trpc.tsx
