+++ cli/template/extras/src/pages/api/auth/[...nextauth]/base.ts
