+++ cli/template/extras/src/pages/api/auth/[...nextauth]/with-prisma.ts
