+++ cli/template/extras/src/pages/api/trpc/[trpc].ts
