+++ cli/template/extras/src/pages/index/with-auth-trpc.tsx
