+++ cli/template/extras/src/pages/index/with-trpc-tw.tsx
