+++ cli/template/extras/src/pages/index/with-tw.tsx
