+++ cli/template/extras/src/server/api/routers/example/base.ts
@@ -1,0 +1,13 @@
+import { z } from "zod";
+
+import { createTrpcRouter, publicProcedure } from "../trpc";
+
+export const exampleRouter = createTrpcRouter({
+  hello: publicProcedure
+    .input(z.object({ text: z.string() }))
+    .query(({ input }) => {
+      return {
+        greeting: `Hello ${input.text}`,
+      };
+    }),
+});
