+++ cli/template/extras/src/server/api/routers/example/with-auth-prisma.ts
@@ -1,0 +1,25 @@
+import { z } from "zod";
+
+import { 
+  createTrpcRouter, 
+  publicProcedure, 
+  protectedProcedure
+} from "../trpc";
+
+export const exampleRouter = createTrpcRouter({
+  hello: publicProcedure
+    .input(z.object({ text: z.string() }))
+    .query(({ input }) => {
+      return {
+        greeting: `Hello ${input.text}`,
+      };
+    }),
+
+  getAll: publicProcedure.query(({ ctx }) => {
+    return ctx.prisma.example.findMany();
+  }),
+
+  getSecretMessage: protectedProcedure.query(() => {
+    return "you can now see this secret message!";
+  }),
+});
