+++ cli/template/extras/src/styles/globals.css
