+++ cli/template/extras/src/types/next-auth.d.ts
